update status string to in

+++ UserService.API/Extensions/ApiResponseExtensions.cs
@@ -18,7 +18,7 @@ namespace UserService.API.Extensions
                 return controller.StatusCode((int)HttpStatusCode.InternalServerError, err);
             }
 
-            if (response.StatusCode == HttpStatusCode.NoContent)
+            if (response.StatusCode == (int)HttpStatusCode.NoContent)
                 return controller.NoContent();
 
             return controller.StatusCode((int)response.StatusCode, response);
